rename constructor parameters
